chore: rename roleGroup to page
