+++ src/app/(dashboard)/[guildUrlName]/page.tsx
@@ -1,10 +1,10 @@
 import type { DynamicRoute } from "@/lib/types";
-import GuildPage from "./[roleGroupUrlName]/page";
+import GuildPage from "./[pageUrlName]/page";
 
 const DefaultGuildPage = async ({
   params,
 }: DynamicRoute<{ guildUrlName: string }>) => {
-  return <GuildPage params={{ ...(await params), roleGroupUrlName: "" }} />;
+  return <GuildPage params={{ ...(await params), pageUrlName: "" }} />;
 };
 
 export default DefaultGuildPage;
